Move SearchClient pytest fixture
By moving it "up" the hierarchy it's automatically available to
all tests. The only reason it manually gets imported is for type annotations.

This is being done to avoid a future circular import issue between get_expr_data/conftest.py
and search/conftest.py
