+++ cegs_portal/search/views/v1/tests/test_target_reg_effects.py
@@ -4,7 +4,7 @@ from typing import cast
 import pytest
 from django.test import Client
 
-from cegs_portal.search.conftest import SearchClient
+from cegs_portal.conftest import SearchClient
 from cegs_portal.search.models import DNAFeature
 
 pytestmark = pytest.mark.django_db
